Rebases to main and integrates the configuration
The commit implements the integration of the YAML configuration into
the Let's Encrypt module. As it requires that the domain and email
address to be set before installation, the corresponding logic was
implemented.

Besides this, the commit creates custom exception classes that are
caught in the Main module and used to return an error message.

Resolves: #9

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>