Implements the execution on multiple hosts
The commit adapts the code for executing operations on multiple
target hosts in the same command, based on a list specified by a new
argument.

As the result consisted in a single value, there were changes in the
return type of the operations (tuple by now, but will be migrated to
a structure in a next commit).

Resolves: #16

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>