Adds support for Let's Encrypt
* Adds Let's Encrypt implementation

This commit adds the Let's Encrypt automation using Certbot. At the
moment it is hard coded for our testing domains.

This is the first iteration of the implementation. In the next commit
there will be added the feature with YAML file for custom email and
domains.

Resolves: #9

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Rebases to main and integrates the configuration

The commit implements the integration of the YAML configuration into
the Let's Encrypt module. As it requires that the domain and email
address to be set before installation, the corresponding logic was
implemented.

Besides this, the commit creates custom exception classes that are
caught in the Main module and used to return an error message.

Resolves: #9

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Fixes a configuration issue

Fixes a configuration issue that overwrites a new file with an empty
one. Besides this, it solves another typo that made the
uninstallation fail.

Resolves: #9

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Fixes bugs on testing and configuration setting

Fixes a bug in testing, produced by a wrong date string formatting.
Another issue that was identified by testing and fixed was a
misbehavior when changing the domain to another.

Resolves: #9

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Changes the result returning in solutions

After rebasing with main, there were present two different approaches
of returning a result from a solution. The commit uniforms this
mechanism to avoid issues in result handling by the CLI component.

Resolves: #9

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

Co-authored-by: iosifache <georgeiosif@live.com>