Refactors and test the API for solutions
The commit refactors the module defining the API used to integrate
security solutions in MutableSecurity. It also tests the logic via
pytest. The solutions implementations were not adapted to the new API
because this will be included in a future commit.

Besides this, the commit achieves the following sub-tasks:
- Adds coverage with coverage.py (horray, a score greater than 90%);
- Adds a new spellchecker in pylint to deal with errors in docstrings
(that are not processed by flake8-spellcheck); and
- Renames the codebase folder.

Resolves: #24

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>