Adds parameters verification for actions
The commit moves the types defined in the information context into a
separate module to let the actions use it. The last mentioned now verify
the values provided by the users and convert it to its real type.

In addition, while adding some exceptions, I observed that some of
them did not have a valid name, so they were renamed and a check was
introduced in the MutableException class.

Resolves: #24

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>