Adapts the teler module for the new API
The commit adapts the existent teler module for the new solutions
API. After testing its functionality, some issues were found and
fixed (especially in the test and information objects).

It adds a new verification for a solution to not be installed on the
system (useful, for example, on installation, to not overwrites its
configuration).

The gevent package, that is used by pyinfra, was modified to suppress
its behavior on errors (printing on stdout/stderr).

Resolves: #24

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>