Implements the auto-generation of documentation
Implements the code for generating automatically the documentation
for solutions integration. The output will be further be taken by a
GitHub action and placed into the future web-based documentation.

As some code from the CLI module was required in the autodoc one, the
functionality was moved into a visual_proxy module to allow the code
sharing.

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>