Automatically generates a new spec sheet for each solution (#40)
* Creates a documentation generation module

Moves functionality from the script generating README.md into a new
module, autodoc, that will help generating Markdown spec sheets too.

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Implements the auto-generation of documentation

Implements the code for generating automatically the documentation
for solutions integration. The output will be further be taken by a
GitHub action and placed into the future web-based documentation.

As some code from the CLI module was required in the autodoc one, the
functionality was moved into a visual_proxy module to allow the code
sharing.

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Generates a solution index

Generates a JSON solution index, from the details already extracted
to generate the spec sheets, to be push into website's repository.

In addition, it changes the unit testing workflow to run only when
changes occurs in tested folders (with Python 3 code inside).

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Adds new workflows and outputs

Adds the workflows for generating the spec sheets and triggering a
workflow from website's repository.

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Renames the word "status"

Only replaces occurances of "status" with "maturity".

Resolves: #39

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>