Adds support for ClamAV (#55)
* Implements new Let's Encrypt x Certbot

This commit implements the refactored version of Let's Encrypt x
Certbot for the new MutableSecurity structure.

Resolves: #51

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Improves and fixes Let's Encrypt x Cerbot

This commit fixes the issues flagged by George-Andrei Iosif
<georgeiosif@live.com>. There is one more issue left. The issue
represents the generation and revocation of certificates when the
user changes the domain or the email.

Resolves: #51

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Implements new ClamAV

This commit implements the refactored version of ClamAV for the
new MutableSecurity structure. Furthermore, there have been a few
bug fixes to the other solutions, as well as adding new categories
for solutions. Those categories are ANTIMALWARE and HOST_PROTECTION.

Resolves: #54

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Minor Fixes to ClamAV

This commit fixes the description of ClamAV, the description of the
test_scan, added the freshclam and daemon description to the ClamAV
description, the typo from scan_day_of_week, added an exit code for
the scan action, deleted the tab from the files.py file for
Deepsource to not fail anymore.

Resolves: #54

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Removed the daemon service

In this commit the daemon service is removed from ClamAV, the
reason behind this is the fact that it was redundant for the stage
of this implementation.

Resolves: #54

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Fixes redundant code and pyinfra functions

This commit fixes two of the SET_INFORMATION functions to be able
to run everything more smoothly and without problems.

Resolves: #54

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Removes Let's Encrypt commit from ClamAV

This commit removes the Let's Encrypt commit from ClamAV to be able
to implement them separately.

Resolves: #54

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>

* Solves formatting issues

Apply auto-formatting and solves warnings/errors raised by Flake8 and
pylint.

Resolves: #54

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Changes text shown in UIs

Changes the formatting of texts shown in user interfaces (both CLI
and documentation).

Resolves: #54

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

* Make log retrieval command not rise error

Make the log retrieval command not raise errors if one of the cat-ed
files not exists.

Resolves: #54

Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>

Signed-off-by: AntociAlin <antoci.alin97@gmail.com>
Signed-off-by: George-Andrei Iosif <georgeiosif@live.com>
Co-authored-by: George-Andrei Iosif <georgeiosif@live.com>